Add files via upload
The Music Recommendation System is an open-source machine learning project hosted on GitHub that aims to provide personalized music recommendations to users based on their music preferences and listening history. This project leverages state-of-the-art machine learning techniques and a large dataset of music tracks to deliver accurate and enjoyable music suggestions.